resume_matcher_app/backend/requirements.txt
Flask
pdfplumber
google-generativeai
werkzeug
streamlit
requests
python-dotenv
